+++ src/ability/ability.factory/ability.factory.spec.ts
@@ -1,0 +1,7 @@
+import { AbilityFactory } from "./ability.factory";
+
+describe("AbilityFactory", () => {
+    it("should be defined", () => {
+        expect(new AbilityFactory()).toBeDefined();
+    });
+});