+++ src/features/features.controller.ts
@@ -34,8 +34,6 @@ import {
     ExtendedFeaturesResponse,
     FeatureResponse,
 } from "./features.response";
-import { AppPermissions } from "../auth/auth.permissions";
-import { Permissions } from "../global/decorators/permissions.decorator";
 import { CustomRequest } from "../global/types/CustomRequest";
 
 @Controller()
@@ -67,7 +65,7 @@ export class FeaturesController {
         description: "Successfully created a new feature.",
         type: FeatureResponse,
     })
-    @Permissions(AppPermissions.OWN_TEAM)
+    // TODO: Add own_team permission here
     @Post("/teams/:teamId/features")
     @ApiCreatedResponse({ type: Feature })
     async createFeature(
@@ -119,7 +117,7 @@ export class FeaturesController {
             "Could not find features for project. Team with given ID does not exist.",
         type: NotFoundErrorResponse,
     })
-    @Permissions(AppPermissions.OWN_TEAM)
+    // TODO: Add own_team permission here
     @Get("/teams/:teamId/features")
     findAllFeatures(@Param("teamId", ParseIntPipe) teamId: number) {
         return this.featuresService.findAllFeatures(teamId);