+++ src/ideations/ideations.service.ts
@@ -1,15 +1,16 @@
 import {
     BadRequestException,
     ConflictException,
+    ForbiddenException,
     Injectable,
     NotFoundException,
-    UnauthorizedException,
 } from "@nestjs/common";
 import { PrismaService } from "../prisma/prisma.service";
 import { CreateIdeationDto } from "./dto/create-ideation.dto";
 import { UpdateIdeationDto } from "./dto/update-ideation.dto";
 import { GlobalService } from "../global/global.service";
 import { CustomRequest } from "../global/types/CustomRequest";
+import { manageOwnVoyageTeamWithIdParam } from "../ability/conditions/voyage-teams.ability";
 
 @Injectable()
 export class IdeationsService {
@@ -19,13 +20,15 @@ export class IdeationsService {
     ) {}
 
     async createIdeation(
-        req,
+        req: CustomRequest,
         teamId: number,
         createIdeationDto: CreateIdeationDto,
     ) {
         const { title, description, vision } = createIdeationDto;
 
         try {
+            // user can create Ideation for their own team(s)
+            manageOwnVoyageTeamWithIdParam(req.user, teamId);
             const createdIdeation = await this.prisma.projectIdea.create({
                 data: {
                     voyageTeamMemberId:
@@ -44,6 +47,8 @@ export class IdeationsService {
     }
 
     async createIdeationVote(req, teamId: number, ideationId: number) {
+        // user can create Ideation votes for their own team(s)
+        manageOwnVoyageTeamWithIdParam(req.user, teamId);
         const ideationExistsCheck = await this.prisma.projectIdea.findUnique({
             where: {
                 id: ideationId,
@@ -83,7 +88,9 @@ export class IdeationsService {
         }
     }
 
-    async getIdeationsByVoyageTeam(id: number) {
+    async getIdeationsByVoyageTeam(req: CustomRequest, id: number) {
+        // user can read Ideation votes for their own team(s)
+        manageOwnVoyageTeamWithIdParam(req.user, id);
         const teamMemberCheck = await this.prisma.voyageTeamMember.findFirst({
             where: {
                 voyageTeamId: id,
@@ -192,8 +199,8 @@ export class IdeationsService {
                 });
                 return updatedIdeation;
             } else {
-                throw new UnauthorizedException(
-                    "You can only update your own project ideas.",
+                throw new ForbiddenException(
+                    "[Ideation Service]:  You can only update your own project ideas.",
                 );
             }
         } catch (e) {
@@ -287,6 +294,7 @@ export class IdeationsService {
         teamId: number,
         ideationId: number,
     ) {
+        manageOwnVoyageTeamWithIdParam(req.user, teamId);
         try {
             const currentSelection = await this.getSelectedIdeation(teamId);
             if (currentSelection) {
@@ -331,23 +339,6 @@ export class IdeationsService {
         throw new NotFoundException(`no ideation found for team ${teamId}`);
     }
 
-    // TODO: this function seems to be unused but might be useful for making new permission guard
-    private async getTeamMemberIdByIdeation(ideationId: number) {
-        const voyageTeamMemberId = await this.prisma.projectIdea.findFirst({
-            where: {
-                id: ideationId,
-            },
-            select: {
-                voyageTeamMemberId: true,
-            },
-        });
-        if (!voyageTeamMemberId)
-            throw new NotFoundException(
-                `Ideation (id: ${ideationId}) does not exist`,
-            );
-        return voyageTeamMemberId;
-    }
-
     private async hasIdeationVote(teamMemberId: number, ideationId: number) {
         const checkVoteStatus = await this.prisma.projectIdeaVote.findMany({
             where: {
@@ -360,7 +351,7 @@ export class IdeationsService {
         });
         if (!checkVoteStatus)
             throw new BadRequestException(
-                "Invalid Ideation Id or Team Member Id",
+                `Invalid Ideation Id or Team Member Id. The user (teamMemberId:${teamMemberId}) does not have a vote for ideation id: ${ideationId}`,
             );
         return checkVoteStatus.length > 0;
     }
@@ -396,7 +387,7 @@ export class IdeationsService {
         });
         if (!oneIdeationVote)
             throw new BadRequestException(
-                "Invalid Ideation Id or Team Member Id",
+                `Invalid Ideation Id or Team Member Id. The user (teamMemberId:${voyageTeamMemberId}) does not have a vote for ideation id: ${ideationId}`,
             );
         return oneIdeationVote;
     }