+++ CHANGELOG.md
@@ -30,9 +30,8 @@ Another example [here](https://co-pilot.dev/changelog)
 - Add multiple device support ([#128](https://github.com/chingu-x/chingu-dashboard-be/pull/128))
 - Add voyage project submission form seed ([#131](https://github.com/chingu-x/chingu-dashboard-be/pull/131))
 - Add voyage project submission controller, service, e2e tests, responses seed ([#133](https://github.com/chingu-x/chingu-dashboard-be/pull/133))
-
-
 - Add new endpoints to select/reset team project ideation ([#136](https://github.com/chingu-x/chingu-dashboard-be/pull/136))
+- Add CASL ability for Access control ([#141](https://github.com/chingu-x/chingu-dashboard-be/pull/141))
 
 ### Changed
 
@@ -64,3 +63,4 @@ Another example [here](https://co-pilot.dev/changelog)
 ### Removed
 
 - Removed email from reset-password request body ([#118](https://github.com/chingu-x/chingu-dashboard-be/pull/118))
+- Removed Roles and Permission guards and decorators([#141](https://github.com/chingu-x/chingu-dashboard-be/pull/141))